+++ Octokit.Reactive/Clients/IObservableTeamsClient.cs
@@ -22,7 +22,7 @@ namespace Octokit.Reactive
         /// <returns>The <see cref="Team"/> with the given identifier.</returns>
         [SuppressMessage("Microsoft.Naming", "CA1716:IdentifiersShouldNotMatchKeywords", MessageId = "Get",
             Justification = "Method makes a network request")]
-        IObservable<Team> Get(int id);
+        IObservable<Team> Get(long id);
 
         /// <summary>
         /// Returns all <see cref="Team" />s for the current org.
@@ -65,7 +65,7 @@ namespace Octokit.Reactive
         /// https://developer.github.com/v3/orgs/teams/#list-child-teams
         /// </remarks>
         /// <param name="id">The team identifier</param>
-        IObservable<Team> GetAllChildTeams(int id);
+        IObservable<Team> GetAllChildTeams(long id);
 
         /// <summary>
         /// Returns all child teams of the given team.
@@ -75,26 +75,26 @@ namespace Octokit.Reactive
         /// </remarks>
         /// <param name="id">The team identifier</param>
         /// <param name="options">Options for changing the API response</param>
-        IObservable<Team> GetAllChildTeams(int id, ApiOptions options);
+        IObservable<Team> GetAllChildTeams(long id, ApiOptions options);
 
         /// <summary>
-        /// Returns all members of the given team. 
+        /// Returns all members of the given team.
         /// </summary>
         /// <remarks>
         /// https://developer.github.com/v3/orgs/teams/#list-team-members
         /// </remarks>
         /// <param name="id">The team identifier</param>
-        IObservable<User> GetAllMembers(int id);
+        IObservable<User> GetAllMembers(long id);
 
         /// <summary>
-        /// Returns all members of the given team. 
+        /// Returns all members of the given team.
         /// </summary>
         /// <remarks>
         /// https://developer.github.com/v3/orgs/teams/#list-team-members
         /// </remarks>
         /// <param name="id">The team identifier</param>
         /// <param name="options">Options to change API behaviour.</param>
-        IObservable<User> GetAllMembers(int id, ApiOptions options);
+        IObservable<User> GetAllMembers(long id, ApiOptions options);
 
         /// <summary>
         /// Returns all members with the specified role in the given team of the given role.
@@ -104,7 +104,7 @@ namespace Octokit.Reactive
         /// </remarks>
         /// <param name="id">The team identifier</param>
         /// <param name="request">The request filter</param>
-        IObservable<User> GetAllMembers(int id, TeamMembersRequest request);
+        IObservable<User> GetAllMembers(long id, TeamMembersRequest request);
 
         /// <summary>
         /// Returns all members with the specified role in the given team of the given role.
@@ -115,7 +115,7 @@ namespace Octokit.Reactive
         /// <param name="id">The team identifier</param>
         /// <param name="request">The request filter</param>
         /// <param name="options">Options to change API behaviour.</param>
-        IObservable<User> GetAllMembers(int id, TeamMembersRequest request, ApiOptions options);
+        IObservable<User> GetAllMembers(long id, TeamMembersRequest request, ApiOptions options);
 
         /// <summary>
         /// Returns newly created <see cref="Team" /> for the current org.
@@ -129,7 +129,7 @@ namespace Octokit.Reactive
         /// To edit a team, the authenticated user must either be an organization owner or a team maintainer
         /// </summary>
         /// <remarks>
-        /// See the <a href="https://docs.github.com/en/rest/teams/teams?apiVersion=2022-11-28#update-a-team">API documentation</a> 
+        /// See the <a href="https://docs.github.com/en/rest/teams/teams?apiVersion=2022-11-28#update-a-team">API documentation</a>
         /// for more information.
         /// </remarks>
         /// <returns>updated <see cref="Team" /> for the current org</returns>
@@ -139,22 +139,22 @@ namespace Octokit.Reactive
         /// Returns updated <see cref="Team" /> for the current org.
         /// This endpoint route is deprecated and will be removed from the Teams API.
         /// We recommend migrating your existing code to use the new Update a team endpoint.
-        /// <see cref="Update(string, string, UpdateTeam)"/>
+        /// <see cref="Update(long, UpdateTeam)"/>
         /// </summary>
         /// <remarks>
-        /// See the <a href="https://docs.github.com/en/rest/teams/teams?apiVersion=2022-11-28#update-a-team-legacy">API documentation</a> 
+        /// See the <a href="https://docs.github.com/en/rest/teams/teams?apiVersion=2022-11-28#update-a-team-legacy">API documentation</a>
         /// for more information.
         /// </remarks>
         /// <exception cref="ApiException">Thrown when a general API error occurs.</exception>
         /// <returns>Updated <see cref="Team"/></returns>
-        IObservable<Team> Update(int id, UpdateTeam team);
+        IObservable<Team> Update(long id, UpdateTeam team);
 
         /// <summary>
         /// To delete a team, the authenticated user must be an organization owner or team maintainer.
         /// If you are an organization owner, deleting a parent team will delete all of its child teams as well.
         /// </summary>
         /// <remarks>
-        /// See the <a href="https://docs.github.com/en/rest/teams/teams?apiVersion=2022-11-28#delete-a-team">API documentation</a> 
+        /// See the <a href="https://docs.github.com/en/rest/teams/teams?apiVersion=2022-11-28#delete-a-team">API documentation</a>
         /// </remarks>
         /// <param name="org">The organization name. The name is not case sensitive.</param>
         /// <param name="teamSlug">The slug of the team name.</param>
@@ -169,12 +169,12 @@ namespace Octokit.Reactive
         /// <see cref="Delete(string, string)"/>.
         /// </summary>
         /// <remarks>
-        /// See the <a href="https://docs.github.com/en/rest/teams/teams?apiVersion=2022-11-28#delete-a-team-legacy">API documentation</a> 
+        /// See the <a href="https://docs.github.com/en/rest/teams/teams?apiVersion=2022-11-28#delete-a-team-legacy">API documentation</a>
         /// </remarks>
         /// <param name="id">The unique identifier of the team.</param>
         /// <exception cref="ApiException">Thrown when a general API error occurs.</exception>
         /// <returns></returns>
-        IObservable<Unit> Delete(int id);
+        IObservable<Unit> Delete(long id);
 
         /// <summary>
         /// Adds a <see cref="User"/> to a <see cref="Team"/>.
@@ -185,7 +185,7 @@ namespace Octokit.Reactive
         /// <param name="id">The team identifier.</param>
         /// <param name="login">The user to add to the team.</param>
         /// <param name="request">Additional parameters for the request</param>
-        IObservable<TeamMembershipDetails> AddOrEditMembership(int id, string login, UpdateTeamMembership request);
+        IObservable<TeamMembershipDetails> AddOrEditMembership(long id, string login, UpdateTeamMembership request);
 
         /// <summary>
         /// Removes a <see cref="User"/> from a <see cref="Team"/>.
@@ -196,10 +196,10 @@ namespace Octokit.Reactive
         /// <param name="id">The team identifier.</param>
         /// <param name="login">The user to remove from the team.</param>
         /// <returns><see langword="true"/> if the user was removed from the team; <see langword="false"/> otherwise.</returns>
-        IObservable<bool> RemoveMembership(int id, string login);
+        IObservable<bool> RemoveMembership(long id, string login);
 
         /// <summary>
-        /// Gets whether the user with the given <paramref name="login"/> 
+        /// Gets whether the user with the given <paramref name="login"/>
         /// is a member of the team with the given <paramref name="id"/>.
         /// A <see cref="NotFoundException"/> is thrown if the user is not a member.
         /// </summary>
@@ -208,7 +208,7 @@ namespace Octokit.Reactive
         /// </remarks>
         /// <param name="id">The team to check.</param>
         /// <param name="login">The user to check.</param>
-        IObservable<TeamMembershipDetails> GetMembershipDetails(int id, string login);
+        IObservable<TeamMembershipDetails> GetMembershipDetails(long id, string login);
 
         /// <summary>
         /// Returns all team's repositories.
@@ -216,7 +216,7 @@ namespace Octokit.Reactive
         /// <param name="id">Team Id.</param>
         /// <exception cref="ApiException">Thrown when a general API error occurs.</exception>
         /// <returns>The team's repositories</returns>
-        IObservable<Repository> GetAllRepositories(int id);
+        IObservable<Repository> GetAllRepositories(long id);
 
         /// <summary>
         /// Returns all team's repositories.
@@ -225,14 +225,14 @@ namespace Octokit.Reactive
         /// <param name="options">Options to change API behaviour.</param>
         /// <exception cref="ApiException">Thrown when a general API error occurs.</exception>
         /// <returns>The team's repositories</returns>
-        IObservable<Repository> GetAllRepositories(int id, ApiOptions options);
+        IObservable<Repository> GetAllRepositories(long id, ApiOptions options);
 
         /// <summary>
         /// Remove a repository from the team
         /// </summary>
         /// <exception cref="ApiException">Thrown when a general API error occurs.</exception>
         /// <returns></returns>
-        IObservable<bool> RemoveRepository(int id, string organization, string repoName);
+        IObservable<bool> RemoveRepository(long id, string organization, string repoName);
 
         /// <summary>
         /// Adds a <see cref="Repository"/> to a <see cref="Team"/>.
@@ -245,7 +245,7 @@ namespace Octokit.Reactive
         /// See the <a href="https://developer.github.com/v3/orgs/teams/#add-team-repo">API documentation</a> for more information.
         /// </remarks>
         /// <returns><see langword="true"/> if the repository was added to the team; <see langword="false"/> otherwise.</returns>
-        IObservable<bool> AddRepository(int id, string organization, string repoName);
+        IObservable<bool> AddRepository(long id, string organization, string repoName);
 
         /// <summary>
         /// Adds a <see cref="Repository"/> to a <see cref="Team"/>.
@@ -259,7 +259,7 @@ namespace Octokit.Reactive
         /// See the <a href="https://developer.github.com/v3/orgs/teams/#add-team-repo">API documentation</a> for more information.
         /// </remarks>
         /// <returns><see langword="true"/> if the repository was added to the team; <see langword="false"/> otherwise.</returns>
-        IObservable<bool> AddRepository(int id, string organization, string repoName, RepositoryPermissionRequest permission);
+        IObservable<bool> AddRepository(long id, string organization, string repoName, RepositoryPermissionRequest permission);
 
         /// <summary>
         /// Gets whether or not the given repository is managed by the given team.
@@ -271,7 +271,7 @@ namespace Octokit.Reactive
         /// See the <a href="https://developer.github.com/v3/orgs/teams/#get-team-repo">API documentation</a> for more information.
         /// </remarks>
         /// <returns><see langword="true"/> if the repository is managed by the given team; <see langword="false"/> otherwise.</returns>
-        IObservable<bool> IsRepositoryManagedByTeam(int id, string owner, string repo);
+        IObservable<bool> IsRepositoryManagedByTeam(long id, string owner, string repo);
 
         /// <summary>
         /// List all pending invitations for the given team.
@@ -282,7 +282,7 @@ namespace Octokit.Reactive
         /// </remarks>
         /// <param name="id">The team identifier</param>
         /// <returns></returns>
-        IObservable<OrganizationMembershipInvitation> GetAllPendingInvitations(int id);
+        IObservable<OrganizationMembershipInvitation> GetAllPendingInvitations(long id);
 
         /// <summary>
         /// List all pending invitations for the given team.
@@ -294,7 +294,7 @@ namespace Octokit.Reactive
         /// <param name="id">The team identifier</param>
         /// <param name="options">Options to change API behaviour.</param>
         /// <returns></returns>
-        IObservable<OrganizationMembershipInvitation> GetAllPendingInvitations(int id, ApiOptions options);
+        IObservable<OrganizationMembershipInvitation> GetAllPendingInvitations(long id, ApiOptions options);
 
         /// <summary>
         /// Checks whether a team has admin, push, maintain, triage, or pull permission for a repository.
@@ -339,9 +339,9 @@ namespace Octokit.Reactive
         /// <param name="owner">The account owner of the repository. The name is not case sensitive.</param>
         /// <param name="repo">The name of the repository. The name is not case sensitive.</param>
         /// <param name="permission">
-        /// The permission to grant the team on this repository. We accept the following permissions to be set: 
-        /// pull, triage, push, maintain, admin and you can also specify a custom repository role name, if the 
-        /// owning organization has defined any. If no permission is specified, the team's permission attribute 
+        /// The permission to grant the team on this repository. We accept the following permissions to be set:
+        /// pull, triage, push, maintain, admin and you can also specify a custom repository role name, if the
+        /// owning organization has defined any. If no permission is specified, the team's permission attribute
         /// will be used to determine what permission to grant the team on this repository
         /// </param>
         /// <exception cref="ApiException">Thrown when a general API error occurs.</exception>
@@ -362,7 +362,7 @@ namespace Octokit.Reactive
         /// <exception cref="ApiException">Thrown when a general API error occurs.</exception>
         /// <returns></returns>
         IObservable<Unit> RemoveRepositoryFromATeam(string org, string teamSlug, string owner, string repo);
-        
+
         /// <summary>
         /// Get a team by slug name
         /// </summary>
